requirements.txt
google-generativeai
sentence-transformers
chromadb
spacy
streamlit
pandas
numpy